Extract out `PhysicsEntityDiagnostics` and improve docs
