+++ src/diagnostics/mod.rs
@@ -1,34 +1,159 @@
-//! Basic diagnostics support.
+//! Diagnostics support for tracking physics timers and counters. Useful for profiling and debugging.
+//!
+//! # Overview
+//!
+//! Each physics plugin such as [`NarrowPhasePlugin`] and [`SolverPlugin`] is responsible
+//! for implementing its own diagnostics resource using the [`PhysicsDiagnostics`] trait
+//! and registering it using [`AppDiagnosticsExt::register_physics_diagnostics`].
+//!
+//! If the `bevy_diagnostic` feature is enabled and the [`PhysicsDiagnosticsPlugin`] is added to the app,
+//! these diagnostics will also be automatically written to the [`Diagnostics`] resource.
+//!
+//! If the `diagnostic_ui` feature is enabled and the [`PhysicsDiagnosticsUiPlugin`] is added to the app,
+//! a debug UI will also be available for displaying these diagnostics in real-time.
+//!
+//! [`NarrowPhasePlugin`]: crate::narrow_phase::NarrowPhasePlugin
+//! [`SolverPlugin`]: crate::dynamics::solver::SolverPlugin
+//! [`PhysicsDiagnosticsUiPlugin`]: crate::diagnostics::ui::PhysicsDiagnosticsUiPlugin
+//!
+//! # Example
+//!
+//! ```no_run
+#![cfg_attr(feature = "2d", doc = "use avian2d::prelude::*;")]
+#![cfg_attr(feature = "3d", doc = "use avian3d::prelude::*;")]
+//! use bevy::prelude::*;
+//!
+//! fn main() {
+//!     App::new()
+//!         .add_plugins((
+//!             DefaultPlugins,
+//!             PhysicsPlugins::default(),
+//!             // Add the `PhysicsDiagnosticsPlugin` to write physics diagnostics
+//!             // to the `DiagnosticsStore` resource in `bevy_diagnostic`.
+//!             // Requires the `bevy_diagnostic` feature.
+//! #           #[cfg(feature = "bevy_diagnostic")]
+//!             PhysicsDiagnosticsPlugin,
+//!             // Add the `PhysicsDiagnosticsUiPlugin` to display physics diagnostics
+//!             // in a debug UI. Requires the `diagnostic_ui` feature.
+//! #           #[cfg(feature = "diagnostic_ui")]
+//!             PhysicsDiagnosticsUiPlugin,
+//!         ))
+//!         // ...your other plugins, systems and resources
+//!         .run();
+//! }
+//! ```
+//!
+//! # Supported Diagnostics
+//!
+//! The following diagnostics are available but not added by default:
+//!
+//! - [`PhysicsTotalDiagnostics`]: Total physics timers (requires the `bevy_diagnostic` feature).
+//! - [`PhysicsEntityDiagnostics`]: Physics entity counters.
+//!
+//! Additionally, physics plugins may implement their own diagnostics that *are* enabled by default.
+//! These include:
+//!
+//! - [`CollisionDiagnostics`]: Diagnostics for collision detection.
+//! - [`SolverDiagnostics`]: Diagnostics for the physics solver.
+//! - [`SpatialQueryDiagnostics`]: Diagnostics for spatial queries.
+//! - [`PhysicsPickingDiagnostics`]: Diagnostics for physics picking.
+//!
+//! [`CollisionDiagnostics`]: crate::collision::CollisionDiagnostics
+//! [`SolverDiagnostics`]: crate::dynamics::solver::SolverDiagnostics
+//! [`SpatialQueryDiagnostics`]: crate::spatial_query::SpatialQueryDiagnostics
+//! [`PhysicsPickingDiagnostics`]: crate::picking::PhysicsPickingDiagnostics
 
+mod entity_counters;
 mod path_macro;
 #[cfg(feature = "bevy_diagnostic")]
 mod total_timers;
 
 #[cfg(feature = "diagnostic_ui")]
 pub mod ui;
+pub use entity_counters::PhysicsEntityDiagnostics;
 pub(crate) use path_macro::impl_diagnostic_paths;
 #[cfg(feature = "bevy_diagnostic")]
 pub use total_timers::{PhysicsTotalDiagnostics, PhysicsTotalDiagnosticsPlugin};
 
-use crate::{
-    collision::ColliderMarker,
-    dynamics::{rigid_body::RigidBody, solver::joints::*},
-    schedule::PhysicsSchedule,
-    PhysicsStepSet,
-};
+use crate::{schedule::PhysicsSchedule, PhysicsStepSet};
 use bevy::{
     app::{App, Plugin},
     diagnostic::{Diagnostic, DiagnosticPath, Diagnostics, RegisterDiagnostic},
-    prelude::{IntoSystemConfigs, IntoSystemSetConfigs, Query, Res, ResMut, Resource, SystemSet},
+    prelude::{IntoSystemConfigs, IntoSystemSetConfigs, Res, ResMut, Resource, SystemSet},
     utils::Duration,
 };
-use bevy::{prelude::ReflectResource, reflect::Reflect};
 
-/// An extension trait for registering physics diagnostics.
+/// A plugin that enables writing [physics diagnostics](crate::diagnostics)
+/// to [`bevy::diagnostic::DiagnosticsStore`]. It is not enabled by default
+/// and must be added manually.
+///
+/// To add a debug UI for physics diagnostics, enable the `diagnostic_ui` feature, and add the
+/// [`PhysicsDiagnosticsUiPlugin`] to your app.
+///
+/// See the [module-level documentation](crate::diagnostics) for more information.
+///
+/// [`PhysicsDiagnosticsUiPlugin`]: crate::diagnostics::ui::PhysicsDiagnosticsUiPlugin
+#[cfg(feature = "bevy_diagnostic")]
+pub struct PhysicsDiagnosticsPlugin;
+
+#[cfg(feature = "bevy_diagnostic")]
+impl Plugin for PhysicsDiagnosticsPlugin {
+    fn build(&self, app: &mut App) {
+        // Configure system sets for physics diagnostics.
+        app.configure_sets(
+            PhysicsSchedule,
+            (
+                PhysicsDiagnosticsSystems::Reset.before(PhysicsStepSet::First),
+                PhysicsDiagnosticsSystems::WriteDiagnostics.after(PhysicsStepSet::Last),
+            ),
+        );
+    }
+}
+
+/// A system set for [physics diagnostics](crate::diagnostics).
+#[derive(SystemSet, Clone, Copy, Debug, PartialEq, Eq, Hash)]
+pub enum PhysicsDiagnosticsSystems {
+    /// Resets diagnostics to their default values.
+    Reset,
+    /// Writes physics diagnostics to other resources, commonly `DiagnosticsStore`.
+    WriteDiagnostics,
+}
+
+/// A trait for resources storing timers and counters for [physics diagnostics](crate::diagnostics).
+pub trait PhysicsDiagnostics: Default + Resource {
+    /// Maps diagnostic paths to their respective duration fields.
+    fn timer_paths(&self) -> Vec<(&'static DiagnosticPath, Duration)> {
+        Vec::new()
+    }
+
+    /// Maps diagnostic paths to their respective counter fields.
+    fn counter_paths(&self) -> Vec<(&'static DiagnosticPath, u32)> {
+        Vec::new()
+    }
+
+    /// A system that resets the diagnostics to their default values.
+    fn reset(mut physics_diagnostics: ResMut<Self>) {
+        *physics_diagnostics = Self::default();
+    }
+
+    /// A system that writes diagnostics to the given [`Diagnostics`] instance.
+    #[cfg(feature = "bevy_diagnostic")]
+    fn write_diagnostics(physics_diagnostics: Res<Self>, mut diagnostics: Diagnostics) {
+        for (path, duration) in physics_diagnostics.timer_paths() {
+            diagnostics.add_measurement(path, || duration.as_secs_f64() * 1000.0);
+        }
+
+        for (path, count) in physics_diagnostics.counter_paths() {
+            diagnostics.add_measurement(path, || count as f64);
+        }
+    }
+}
+
+/// An extension trait for registering [physics diagnostics](crate::diagnostics) in an [`App`].
 pub trait AppDiagnosticsExt {
     /// Registers timer and counter diagnostics for a resource implementing [`PhysicsDiagnostics`].
     ///
-    /// This method should be called in [`Plugin::finish`] to ensure that [`Diagnostics`]
+    /// This method should be called in [`Plugin::finish`] to ensure that [`Diagnostic`]s
     /// are only tracked if the [`PhysicsDiagnosticsPlugin`] was added to the app.
     fn register_physics_diagnostics<T: PhysicsDiagnostics>(&mut self);
 }
@@ -70,10 +195,11 @@ impl AppDiagnosticsExt for App {
             }
 
             for path in counter_paths.iter().map(|(path, _)| *path) {
+                // All counters are in whole numbers.
                 self.register_diagnostic(Diagnostic::new(path.clone()).with_smoothing_factor(0.0));
             }
 
-            // Add systems to reset the diagnostics and write them to the `Diagnostics` resource.
+            // Add systems to reset the diagnostics and write them to the `DiagnosticsStore` resource.
             self.add_systems(
                 PhysicsSchedule,
                 T::write_diagnostics
@@ -81,148 +207,5 @@ impl AppDiagnosticsExt for App {
                     .ambiguous_with_all(),
             );
         }
-    }
-}
-
-/// A system set for physics diagnostics.
-#[derive(SystemSet, Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum PhysicsDiagnosticsSystems {
-    /// Resets diagnostics to their default values.
-    Reset,
-    /// Writes physics diagnostics to other resources, commonly `DiagnosticsStore`.
-    WriteDiagnostics,
-}
-
-/// A trait for resources storing timers and counters for physics diagnostics.
-pub trait PhysicsDiagnostics: Default + Resource {
-    /// Maps diagnostic paths to their respective duration fields.
-    fn timer_paths(&self) -> Vec<(&'static DiagnosticPath, Duration)> {
-        Vec::new()
-    }
-
-    /// Maps diagnostic paths to their respective counter fields.
-    fn counter_paths(&self) -> Vec<(&'static DiagnosticPath, u32)> {
-        Vec::new()
-    }
-
-    /// A system that resets the diagnostics to their default values.
-    fn reset(mut physics_diagnostics: ResMut<Self>) {
-        *physics_diagnostics = Self::default();
-    }
-
-    /// A system that writes diagnostics to the given [`Diagnostics`] instance.
-    #[cfg(feature = "bevy_diagnostic")]
-    fn write_diagnostics(physics_diagnostics: Res<Self>, mut diagnostics: Diagnostics) {
-        for (path, duration) in physics_diagnostics.timer_paths() {
-            diagnostics.add_measurement(path, || duration.as_secs_f64() * 1000.0);
-        }
-
-        for (path, count) in physics_diagnostics.counter_paths() {
-            diagnostics.add_measurement(path, || count as f64);
-        }
-    }
-}
-
-/// A plugin that adds various diagnostics for debugging purposes.
-/// It is not enabled by default and must be added manually.
-#[cfg(feature = "bevy_diagnostic")]
-pub struct PhysicsDiagnosticsPlugin;
-
-#[cfg(feature = "bevy_diagnostic")]
-impl Plugin for PhysicsDiagnosticsPlugin {
-    fn build(&self, app: &mut App) {
-        app.configure_sets(
-            PhysicsSchedule,
-            (
-                PhysicsDiagnosticsSystems::Reset.before(PhysicsStepSet::First),
-                PhysicsDiagnosticsSystems::WriteDiagnostics.after(PhysicsStepSet::Last),
-            ),
-        );
-
-        app.add_systems(
-            PhysicsSchedule,
-            diagnostic_entity_counts.in_set(PhysicsStepSet::First),
-        );
-    }
-
-    fn finish(&self, app: &mut App) {
-        // Register diagnostics for physics entity counts.
-        app.register_physics_diagnostics::<PhysicsEntityDiagnostics>();
-    }
-}
-
-/// Diagnostics for physics entity counts.
-#[derive(Resource, Debug, Default, Reflect)]
-#[reflect(Resource, Debug)]
-#[cfg(feature = "bevy_diagnostic")]
-pub struct PhysicsEntityDiagnostics {
-    /// The number of dynamic bodies.
-    pub dynamic_body_count: u32,
-    /// The number of kinematic bodies.
-    pub kinematic_body_count: u32,
-    /// The number of static bodies.
-    pub static_body_count: u32,
-    /// The number of colliders.
-    pub collider_count: u32,
-    /// The number of joint.
-    pub joint_count: u32,
-}
-
-#[cfg(feature = "bevy_diagnostic")]
-impl PhysicsDiagnostics for PhysicsEntityDiagnostics {
-    fn counter_paths(&self) -> Vec<(&'static DiagnosticPath, u32)> {
-        vec![
-            (Self::DYNAMIC_BODY_COUNT, self.dynamic_body_count),
-            (Self::KINEMATIC_BODY_COUNT, self.kinematic_body_count),
-            (Self::STATIC_BODY_COUNT, self.static_body_count),
-            (Self::COLLIDER_COUNT, self.collider_count),
-            (Self::JOINT_COUNT, self.joint_count),
-        ]
-    }
-}
-
-#[cfg(feature = "bevy_diagnostic")]
-impl_diagnostic_paths! {
-    impl PhysicsEntityDiagnostics {
-        DYNAMIC_BODY_COUNT: "avian/entity_count/dynamic_bodies",
-        KINEMATIC_BODY_COUNT: "avian/entity_count/kinematic_bodies",
-        STATIC_BODY_COUNT: "avian/entity_count/static_bodies",
-        COLLIDER_COUNT: "avian/entity_count/colliders",
-        JOINT_COUNT: "avian/entity_count/joints",
-    }
-}
-
-// TODO: This is pretty inefficient.
-#[cfg(feature = "bevy_diagnostic")]
-fn diagnostic_entity_counts(
-    rigid_bodies_query: Query<&RigidBody>,
-    colliders_query: Query<&ColliderMarker>,
-    fixed_joint_query: Query<&FixedJoint>,
-    prismatic_joint_query: Query<&PrismaticJoint>,
-    distance_joint_query: Query<&DistanceJoint>,
-    revolute_joint_query: Query<&RevoluteJoint>,
-    #[cfg(feature = "3d")] spherical_joint_query: Query<&SphericalJoint>,
-    mut diagnostics: ResMut<PhysicsEntityDiagnostics>,
-) {
-    diagnostics.dynamic_body_count = rigid_bodies_query
-        .iter()
-        .filter(|rb| rb.is_dynamic())
-        .count() as u32;
-    diagnostics.kinematic_body_count = rigid_bodies_query
-        .iter()
-        .filter(|rb| rb.is_kinematic())
-        .count() as u32;
-    diagnostics.static_body_count = rigid_bodies_query
-        .iter()
-        .filter(|rb| rb.is_static())
-        .count() as u32;
-    diagnostics.collider_count = colliders_query.iter().count() as u32;
-    diagnostics.joint_count = fixed_joint_query.iter().count() as u32
-        + prismatic_joint_query.iter().count() as u32
-        + distance_joint_query.iter().count() as u32
-        + revolute_joint_query.iter().count() as u32;
-    #[cfg(feature = "3d")]
-    {
-        diagnostics.joint_count += spherical_joint_query.iter().count() as u32;
     }
 }
